pyproject.toml
[tool.poetry]
name = "dspy-metric-learning"
version = "0.1.0"
description = "A DSPy metric function learning package"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
dspy-ai = "^2.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[tool.pytest.ini_options]
testpaths = ["tests", "integration_tests"]
python_files = "test_*.py"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
